[Feature 133989813] Auth streamlined
