+++ client/components/Landingpage/LandingPage.jsx
@@ -1,6 +1,20 @@
 import React from 'react';
 import AppBar from 'material-ui/AppBar';
-import AuthWrapper from '../Auth/AuthWrapper';
+import Paper from 'material-ui/Paper';
+import AuthWrapperContainer from '../../containers/Auth/AuthWrapperContainer';
+
+const style = {
+  height: 500,
+  width: 500,
+  margin: 20,
+  textAlign: 'center',
+  display: 'inline-block',
+  position: 'absolute',
+  top: '60%',
+  left: '50%',
+  marginRight: '-50%',
+  transform: 'translate(' + '-50%,' + '-50%' + ')'
+};
 
 const LandingPage = () => {
   return (
@@ -14,7 +28,9 @@ const LandingPage = () => {
         <h3>Start now :) </h3>
       </div>
       <div>
-        <AuthWrapper />
+        <Paper style={style} zDepth={3}>
+          <AuthWrapperContainer />
+        </Paper>
       </div>
 
     </div>
