+++ client/routes/Routes.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import {Router, Route, hashHistory, IndexRoute} from 'react-router';
 import {Provider} from 'react-redux';
 import Main from '../components/Main/Main';
-import AuthWrapper from '../components/Auth/AuthWrapper';
 import DashBoard from '../components/Dashboard/DashBoard';
 import LandingPage from '../components/Landingpage/LandingPage';
 import configureStore from '../store/store';
@@ -16,7 +15,6 @@ const routes = (
       <Route path='/' components={Main}>
         <IndexRoute components={LandingPage} />
       <Route path='/home' components={DashBoard} />
-      <Route path='/trial' components={AuthWrapper} />
       </Route>
     </Router>
   </Provider>
