+++ package.json
@@ -8,7 +8,7 @@
     "build": "webpack --config webpack.config.prod.js",
     "dev": "webpack-dev-server --config webpack.config.dev.js",
     "test": "export NODE_ENV=testing; node seed/dropdb.js && node seed/index.js; istanbul cover _mocha tests/* && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
-    "test:fend": "mocha --compilers js:babel-core/register \"tests/client/**/*@(.js|.jsx)\"",
+    "test:fend": "mocha --compilers js:babel-core/register --require ./tests/test-helper.js \"tests/client/**/*@(.js|.jsx)\"",
     "lint": "node_modules/.bin/esw webpack.config.* client",
     "lint:watch": "npm run lint -- --watch"
   },
@@ -53,6 +53,7 @@
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
     "bcrypt-nodejs": "0.0.3",
+    "chai-immutable": "^1.6.0",
     "coveralls": "^2.11.12",
     "css-loader": "^0.25.0",
     "dotenv": "^2.0.0",
@@ -63,6 +64,7 @@
     "eventsource-polyfill": "^0.9.6",
     "istanbul": "^0.4.5",
     "jquery": "^3.1.1",
+    "jsdom": "^9.8.3",
     "jsonwebtoken": "^7.1.9",
     "mocha-lcov-reporter": "^1.2.0",
     "redux-logger": "^2.7.4",
