+++ tests/test-helper.js
@@ -1,0 +1,21 @@
+import jsdom from 'jsdom';
+import chai from 'chai';
+import chaiImmutable from 'chai-immutable';
+
+// set-up jsdom
+// create jsdom versions of windows and documents
+const doc = jsdom.jsdom('<!doctype html><html><body></body></html>');
+const win = doc.defaultView;
+
+// put them on global objects
+global.document = doc;
+global.window = win;
+
+// get properties of the window object such as navigator
+Object.keys(window).forEach((key) => {
+  if(!(key in global)) {
+    global[key] = window[key];
+  }
+});
+
+chai.use(chaiImmutable);
